♻️(front) move CourseRunEnrollment
As this component is not a widget anymore we move it inside the new
widget that exclusively uses it, which is SyllabysCourseRunsList.
