+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseRunEnrollment/_styles.scss
