+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseRunEnrollment/index.tsx
@@ -8,8 +8,8 @@ import { User } from 'types/User';
 import { Maybe, Nullable } from 'types/utils';
 import { handle } from 'utils/errors/handle';
 import { HttpError } from 'utils/errors/HttpError';
-import useCourseEnrollment from './hooks/useCourseEnrollment';
-import { CourseRunUnenrollButton } from './components/CourseRunUnenrollmentButton';
+import useCourseEnrollment from 'widgets/SyllabusCourseRunsList/components/CourseRunEnrollment/hooks/useCourseEnrollment';
+import { CourseRunUnenrollButton } from 'widgets/SyllabusCourseRunsList/components/CourseRunEnrollment/CourseRunUnenrollmentButton';
 
 const messages = defineMessages({
   enroll: {
