+++ src/frontend/scss/components/_index.scss
@@ -21,8 +21,6 @@
 @import '../../js/widgets/CourseProductItem/components/SaleTunnelStepValidation/styles';
 @import '../../js/widgets/CourseProductItem/components/StepBreadcrumb/styles';
 @import '../../js/widgets/CourseProductItem/styles';
-@import '../../js/widgets/CourseRunEnrollment/components/CourseRunUnenrollmentButton/styles';
-@import '../../js/widgets/CourseRunEnrollment/styles';
 @import '../../js/widgets/Dashboard/styles';
 @import '../../js/widgets/Dashboard/components/DashboardAvatar/styles';
 @import '../../js/widgets/Dashboard/components/DashboardBox/styles';
@@ -38,6 +36,8 @@
 @import '../../js/widgets/Search/components/SearchFilterGroup/styles';
 @import '../../js/widgets/Search/components/SearchFilterGroupModal/styles';
 @import '../../js/widgets/Search/components/SearchFiltersPane/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseRunEnrollment/CourseRunUnenrollmentButton/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseRunEnrollment/styles';
 @import '../../js/widgets/SyllabusCourseRunsList/components/CourseWishButton/styles';
 @import '../../js/widgets/UserLogin/components/UserMenu/styles';
 @import '../../js/widgets/UserLogin/styles';
